Add workdir option to save checkpoints

+++ configs/wpod.py
@@ -6,7 +6,7 @@ def get_config() -> ConfigDict:
 
   config.vocab_size = 68
   config.logits_via_embedding = True
-  config.dtype = "float32"
+  config.dtype = "bfloat16"
   config.embed_dim = 256
   config.n_heads = 8
   config.n_layers = 6
